Intresting part Get winners
in this commit i added a function which consist of patterns in which a player will win in form of array by using if i match the patterns if their is a match it will return the matching x  or o else return null. i even added a condition to check if winner is selected then it will render winner in text and x or o accordingly if no winner is selected next turn is displayed. when click on any button with existing o or x and after winner is selected handle event returns which doesnt cause overwriting